# Conflicts:
#	app/src/main/java/com/kloudsync/techexcel/bean/MeetingConfig.java
#	app/src/main/java/com/kloudsync/techexcel/dialog/RecordPlayDialog.java
#	app/src/main/java/com/kloudsync/techexcel/dialog/plugin/UserNotesDialog.java
#	app/src/main/java/com/kloudsync/techexcel/frgment/MeetingFragment.java
#	app/src/main/java/com/kloudsync/techexcel/frgment/TeamDocumentsFragment.java
#	app/src/main/java/com/kloudsync/techexcel/help/MeetingKit.java
#	app/src/main/java/com/kloudsync/techexcel/help/NoteViewManager.java
#	app/src/main/java/com/kloudsync/techexcel/help/PopBottomFile.java
#	app/src/main/java/com/kloudsync/techexcel/help/PopBottomMenu.java
#	app/src/main/java/com/kloudsync/techexcel/help/PopMeetingMenu.java
#	app/src/main/java/com/kloudsync/techexcel/help/RecordShareVedioManager.java
#	app/src/main/java/com/kloudsync/techexcel/help/RtcManager.java
#	app/src/main/java/com/kloudsync/techexcel/tool/SocketMessageManager.java
#	app/src/main/java/com/kloudsync/techexcel/ui/DocAndMeetingActivity.java
#	app/src/main/java/com/kloudsync/techexcel/ui/MainActivity.java
#	app/src/main/java/com/kloudsync/techexcel/ui/MainActivityV2.java
#	app/src/main/java/com/ub/service/KloudWebClientManager.java
#	app/src/main/java/com/ub/service/activity/SyncBookActivity.java
#	app/src/main/java/com/ub/service/activity/SyncRoomActivity.java
#	app/src/main/java/com/ub/service/activity/WatchCourseActivity2.java
#	app/src/main/java/com/ub/service/activity/WatchCourseActivity3.java
#	app/src/main/java/com/ub/techexcel/adapter/AgoraCameraAdapter.java
#	app/src/main/java/com/ub/techexcel/adapter/SyncRoomNoteListAdapter.java
#	app/src/main/java/com/ub/techexcel/bean/AgoraMember.java
#	app/src/main/java/com/ub/techexcel/tools/MeetingRecordsDialog.java
#	app/src/main/java/com/ub/techexcel/tools/ServiceInterfaceTools.java
#	app/src/main/res/layout/activity_doc_and_meeting.xml
#	app/src/main/res/layout/watchcourse3.xml

+++ app/src/main/java/com/kloudsync/techexcel/personal/AccountSettingActivity.java
@@ -174,7 +174,7 @@ public class AccountSettingActivity extends Activity {
     }
 
 
-
+    private boolean isAdmin = false;
 
     private void getSchoolContact() {
 
@@ -185,9 +185,10 @@ public class AccountSettingActivity extends Activity {
                 UserInCompany userInCompany = (UserInCompany) object;
                 if (userInCompany.getRole() == 7 || userInCompany.getRole() == 8) {
                     layout_enable_sync.setVisibility(View.VISIBLE);
-
+                    isAdmin = true;
                 } else {
                     layout_enable_sync.setVisibility(View.GONE);
+                    isAdmin = false;
                 }
             }
         });
@@ -704,7 +705,11 @@ public class AccountSettingActivity extends Activity {
 
                         @Override
                         public void fileDeletePhoto() {
-                            deleteCompanyLogo();
+                            if (isAdmin) {
+                                deleteCompanyLogo();
+                            } else {
+                                Toast.makeText(AccountSettingActivity.this, "当前你不是管理员,没有权限修改", Toast.LENGTH_LONG).show();
+                            }
                         }
 
                         @Override
@@ -730,7 +735,11 @@ public class AccountSettingActivity extends Activity {
                     startActivity(intent);
                     break;
                 case R.id.as_tv_save:
-                    updateCompanyInfo();
+                    if (isAdmin) {
+                        updateCompanyInfo();
+                    } else {
+                        Toast.makeText(AccountSettingActivity.this, "当前你不是管理员,没有权限修改", Toast.LENGTH_LONG).show();
+                    }
                     break;
                 case R.id.layout_sync_name:
                     intent = new Intent(getApplicationContext(), SyncRoomNameActivity.class);
